Add tags to SSM paramters to allow Copilt services to read them
By default Copilo service can only read SSM paramters that are tagged wtih their name and environment.  We however have a need for global account wide SSM parameters that are consumable by any Copilot service in that account.  We already have a set of global SSM paramters but overlooked that an explicit policy was required for the service to be able to access the paramter.

This PR fixes this issue by adding a `Copilot-application: __all__` tag for these account wide SSM parameters. There's an associated PR in the platform-tools account that adds the policy to the Copilot/ECS task role.